Moving realm code to its own module
