+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/factory/internal/DatabaseRepositoryFactory.kt
@@ -36,14 +36,6 @@ import io.getstream.chat.android.offline.repository.domain.queryChannels.interna
 import io.getstream.chat.android.offline.repository.domain.reaction.internal.DatabaseReactionRepository
 import io.getstream.chat.android.offline.repository.domain.syncState.internal.DatabaseSyncStateRepository
 import io.getstream.chat.android.offline.repository.domain.user.internal.DatabaseUserRepository
-import io.getstream.chat.android.offline.repository.realm.repository.RealmChannelConfigRepository
-import io.getstream.chat.android.offline.repository.realm.repository.RealmChannelRepository
-import io.getstream.chat.android.offline.repository.realm.repository.RealmMessageRepository
-import io.getstream.chat.android.offline.repository.realm.repository.RealmQueryChannelsRepository
-import io.getstream.chat.android.offline.repository.realm.repository.RealmReactionRepository
-import io.getstream.chat.android.offline.repository.realm.repository.RealmSyncStateRepository
-import io.getstream.chat.android.offline.repository.realm.repository.RealmUserRepository
-import io.realm.kotlin.Realm
 
 private const val DEFAULT_CACHE_SIZE = 100
 
@@ -51,7 +43,6 @@ private const val DEFAULT_CACHE_SIZE = 100
 internal class DatabaseRepositoryFactory(
     private val database: ChatDatabase,
     private val currentUser: User,
-    private val realm: Realm,
 ) : RepositoryFactory {
 
     private var repositoriesCache: MutableMap<Class<out Any>, Any> = mutableMapOf()
@@ -66,8 +57,6 @@ internal class DatabaseRepositoryFactory(
         }
     }
 
-    private fun realmUserRepository(): RealmUserRepository = RealmUserRepository(realm)
-
     private fun roomChannelConfigRepository(): DatabaseChannelConfigRepository {
         val databaseChannelConfigRepository =
             repositoriesCache[ChannelConfigRepository::class.java] as? DatabaseChannelConfigRepository?
@@ -78,10 +67,6 @@ internal class DatabaseRepositoryFactory(
             }
         }
     }
-
-    private fun realmChannelConfigRepository(): RealmChannelConfigRepository = RealmChannelConfigRepository(realm)
-
-    private fun realmChannelRepository(): ChannelRepository = RealmChannelRepository(realm)
 
     private fun roomChannelRepository(
         getUser: suspend (userId: String) -> User,
@@ -107,10 +92,6 @@ internal class DatabaseRepositoryFactory(
             }
         }
     }
-
-    private fun realmQueryChannelsRepository(): RealmQueryChannelsRepository = RealmQueryChannelsRepository(realm)
-
-    private fun realmMessageRepository(): MessageRepository = RealmMessageRepository(realm)
 
     private fun roomMessageRepository(
         getUser: suspend (userId: String) -> User,
@@ -140,8 +121,6 @@ internal class DatabaseRepositoryFactory(
         }
     }
 
-    private fun realmReactionRepository(): RealmReactionRepository = RealmReactionRepository(realm)
-
     private fun roomSyncStateRepository(): DatabaseSyncStateRepository {
         val databaseSyncStateRepository =
             repositoriesCache[SyncStateRepository::class.java] as? DatabaseSyncStateRepository?
@@ -153,28 +132,27 @@ internal class DatabaseRepositoryFactory(
         }
     }
 
-    private fun realmSyncStateRepository(): RealmSyncStateRepository = RealmSyncStateRepository(realm)
 
-    override fun createUserRepository(): UserRepository = realmUserRepository()
+    override fun createUserRepository(): UserRepository = roomUserRepository()
 
-    override fun createChannelConfigRepository(): ChannelConfigRepository = realmChannelConfigRepository()
+    override fun createChannelConfigRepository(): ChannelConfigRepository = roomChannelConfigRepository()
 
     override fun createChannelRepository(
         getUser: suspend (userId: String) -> User,
         getMessage: suspend (messageId: String) -> Message?,
-    ): ChannelRepository = realmChannelRepository()
+    ): ChannelRepository = roomChannelRepository(getUser, getMessage)
 
-    override fun createQueryChannelsRepository(): QueryChannelsRepository = realmQueryChannelsRepository()
+    override fun createQueryChannelsRepository(): QueryChannelsRepository = roomQueryChannelsRepository()
 
     override fun createMessageRepository(
         getUser: suspend (userId: String) -> User,
-    ): MessageRepository = realmMessageRepository()
+    ): MessageRepository = roomMessageRepository(getUser)
 
     override fun createReactionRepository(
         getUser: suspend (userId: String) -> User
-    ): ReactionRepository = realmReactionRepository()
+    ): ReactionRepository = roomReactionRepository(getUser)
 
-    override fun createSyncStateRepository(): SyncStateRepository = realmSyncStateRepository()
+    override fun createSyncStateRepository(): SyncStateRepository = roomSyncStateRepository()
 
     override fun createAttachmentRepository(): AttachmentRepository {
         val databaseAttachmentRepository =
