+++ stream-chat-android-ui-components-sample/build.gradle
@@ -148,6 +148,7 @@ dependencies {
     implementation project(":stream-chat-android-pushprovider-huawei")
     implementation project(":stream-chat-android-pushprovider-xiaomi")
     implementation project(":stream-chat-android-offline")
+    implementation project(":stream-chat-android-offline-realm")
     implementation files('../libraries/external/MiPush_SDK_Client_5_0_6-G_3rd.aar')
 
     implementation Dependencies.androidxActivityKtx
